move LinkedItem into models package
