+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/home/HomeViewModel.kt
@@ -8,6 +8,7 @@ import app.omnivore.omnivore.Constants
 import app.omnivore.omnivore.DatastoreKeys
 import app.omnivore.omnivore.DatastoreRepository
 import app.omnivore.omnivore.graphql.generated.SearchQuery
+import app.omnivore.omnivore.models.LinkedItem
 import com.apollographql.apollo3.ApolloClient
 import com.apollographql.apollo3.api.Optional
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -123,34 +124,3 @@ class HomeViewModel @Inject constructor(
   }
 }
 
-public data class LinkedItem(
-  public val id: String,
-  public val title: String,
-  public val createdAt: Any,
-//  public val savedAt: Any,
-  public val readAt: Any?,
-//  public val updatedAt: Any,
-  public val readingProgress: Double,
-  public val readingProgressAnchor: Int,
-  public val imageURLString: String?,
-//  public val onDeviceImageURLString: String?,
-//  public val documentDirectoryPath: String?,
-  public val pageURLString: String,
-  public val descriptionText: String?,
-  public val publisherURLString: String?,
-//  public val siteName: String?,
-  public val author: String?,
-  public val publishDate: Any?,
-  public val slug: String,
-//  public val isArchived: Boolean,
-//  public val contentReader: String?,
-//  public val originalHtml: String?,
-) {
-  fun publisherDisplayName(): String? {
-    return publisherURLString?.toUri()?.host
-  }
-
-  fun labelsJSONString(): String {
-    return ""
-  }
-}
