+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/home/LinkedItemCard.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import app.omnivore.omnivore.models.LinkedItem
 import coil.compose.rememberAsyncImagePainter
 
 @Composable
