+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/reader/WebReaderContent.kt
@@ -1,6 +1,6 @@
 package app.omnivore.omnivore.ui.reader
 
-import app.omnivore.omnivore.ui.home.LinkedItem
+import app.omnivore.omnivore.models.LinkedItem
 
 enum class WebFont(val displayText: String, val rawValue: String) {
   INTER("Inter", "Inter"),
